Add RedirectTrait.
This allows easily creating custom route classes for redirection
without duplicating code from RedirectRoute. For e.g.

```
class I18nRedirectRoute extends I18nRoute
{
    use RedirectTrait;
}
```
